barebox-pbl: make compatibility explicit
Instead of deriving the value of "COMPATIBLE_MACHINE" from barebox, it
must now be explicitly stated when the pre-bootloader is compatible with
the machine. That is because not every target machine has support for a
pre-bootloader.

This decision also makes the interface more stable, because now BSP
layers need to include this information, independent of whether such
explicit compatibility statements will still be required in the future
or not.