Distribute compatibility info among concrete recipes
Instead of setting the variables PACKAGE_ARCH and COMPATIBLE_MACHINE in
barebox-common.inc, they are now set in the specialized barebox*.inc
files. The reason is that the common values were not so common after
all, so that barebox-utils.inc had to revert what had already been set
in the common file.

Signed-off-by: Dennis Menschel <menschel-d@posteo.de>