Remove obsolete barebox versions
The original motivation for keeping multiple versions of barebox in a
Yocto branch was to give the user the opportunity to stick to a specific
version during one Yocto release cycle.

After following this policy for one Yocto release cycle I've come to the
following conclusion:

- I'm not aware of any user who has actually benefited from this policy.
- Barebox follows a "rolling release" cycle and does not offer special
  long term support releases.
- Nobody else in the Yocto/OpenEmbedded ecosystem follows such a policy.
  Most recipes are either updated as soon as a new upstream version is
  released, or the maintainers stick to a particular version for the
  complete Yocto release cycle.

Therefore, from now on, only the most recent version of barebox will be
provided. This will also simplify maintenance and cherry-picking for
older branches.

Signed-off-by: Dennis Menschel <menschel-d@posteo.de>