layer.conf: add missing LAYERDEPENDS
As it turns out, the LAYERDEPENDS variable seems to have been available
in Yocto/OE at least since 2014, but was not part of the generated
output of "yocto-layer create" when creating the skeleton for
meta-barebox in 2016 (e9e653cca3314f4ba8b418bd098f9ac8711f9841).
The replacement command "bitbake-layers create-layer" now automatically
creates this LAYERDEPENDS-entry.

Signed-off-by: Dennis Menschel <menschel-d@posteo.de>
Co-authored-by: Jan-Simon Möller <dl9pf@gmx.de>